Fixed some tests, undid MySQL work for now to merge recent improvements in first
