+++ app/connectors_service/connectors/config.py
@@ -124,9 +124,9 @@ def _default_config():
             "jira": "connectors.sources.jira:JiraDataSource",
             "microsoft_teams": "connectors.sources.microsoft_teams:MicrosoftTeamsDataSource",
             "mongodb": "connectors.sources.mongo.datasource:MongoDataSource",
-            "mssql": "connectors.sources.mssql.datasource:MSSQLDataSource",
-            "mysql": "connectors.sources.mysql.datasource:MySqlDataSource",
-            "network_drive": "connectors.sources.network_drive:NASDataSource",
+            "mssql": "connectors.sources.mssql:MSSQLDataSource",
+            "mysql": "connectors.sources.mysql:MySqlDataSource",
+            "network_drive": "connectors.sources.network_drive.datasource:NASDataSource",
             "notion": "connectors.sources.notion:NotionDataSource",
             "onedrive": "connectors.sources.onedrive:OneDriveDataSource",
             "oracle": "connectors.sources.oracle:OracleDataSource",
