+++ app/connectors_service/tests/sources/test_dropbox.py
@@ -22,15 +22,19 @@ from connectors_sdk.filtering.validation import Filter, SyncRuleValidationResult
 from connectors_sdk.source import ConfigurableFieldValueError
 from freezegun import freeze_time
 
-from app.connectors_service.connectors.sources.dropbox.dropbox_client import (
-    AUTHENTICATED_ADMIN_URL,
+from connectors.sources.dropbox.dropbox_client import (
     DropBoxAdvancedRulesValidator,
     DropboxClient,
-    DropboxDataSource,
+)
+from connectors.sources.dropbox.common import (
+    AUTHENTICATED_ADMIN_URL,
     InvalidClientCredentialException,
     InvalidPathException,
-    InvalidRefreshTokenException,
+    InvalidRefreshTokenException
 )
+
+from connectors.sources.dropbox.datasource import DropboxDataSource
+
 from tests.commons import AsyncIterator
 from tests.sources.support import create_source
 
@@ -748,7 +752,7 @@ async def test_ping():
 
 
 @pytest.mark.asyncio
-@patch("connectors.sources.dropbox.RETRY_INTERVAL", 0)
+@patch("connectors.sources.dropbox.common.RETRY_INTERVAL", 0)
 async def test_ping_when_server_timeout_error_raises():
     async with create_source(DropboxDataSource) as source:
         setup_dropbox(source)
@@ -761,7 +765,7 @@ async def test_ping_when_server_timeout_error_raises():
 
 
 @pytest.mark.asyncio
-@patch("connectors.sources.dropbox.RETRY_INTERVAL", 0)
+@patch("connectors.sources.dropbox.common.RETRY_INTERVAL", 0)
 async def test_ping_when_client_response_error_occurs():
     async with create_source(DropboxDataSource) as source:
         setup_dropbox(source)
@@ -783,7 +787,7 @@ async def test_ping_when_client_response_error_occurs():
 
 
 @pytest.mark.asyncio
-@patch("connectors.sources.dropbox.RETRY_INTERVAL", 0)
+@patch("connectors.sources.dropbox.common.RETRY_INTERVAL", 0)
 async def test_ping_when_client_response_error_occur_with_unexpected_url():
     async with create_source(DropboxDataSource) as source:
         setup_dropbox(source)
@@ -805,7 +809,7 @@ async def test_ping_when_client_response_error_occur_with_unexpected_url():
 
 
 @pytest.mark.asyncio
-@patch("connectors.sources.dropbox.RETRY_INTERVAL", 0)
+@patch("connectors.sources.dropbox.common.RETRY_INTERVAL", 0)
 async def test_api_call_negative():
     async with create_source(DropboxDataSource) as source:
         setup_dropbox(source)
@@ -909,7 +913,7 @@ async def test_set_access_token_when_token_expires_at_is_str():
 
 @pytest.fixture
 def patch_default_wait_multiplier():
-    with mock.patch("connectors.sources.dropbox.RETRY_INTERVAL", 0):
+    with mock.patch("connectors.sources.dropbox.common.RETRY_INTERVAL", 0):
         yield
 
 
@@ -980,7 +984,7 @@ async def test_api_call_when_status_429_exception():
 
 
 @pytest.mark.asyncio
-@patch("connectors.sources.dropbox.DEFAULT_RETRY_AFTER", 0)
+@patch("connectors.sources.dropbox.common.DEFAULT_RETRY_AFTER", 0)
 async def test_api_call_when_status_429_exception_without_retry_after_header():
     async with create_source(DropboxDataSource) as source:
         setup_dropbox(source)
