+++ app/connectors_service/tests/sources/test_mysql.py
@@ -14,7 +14,7 @@ from connectors_sdk.logger import logger
 from connectors_sdk.source import ConfigurableFieldValueError
 from freezegun import freeze_time
 
-from app.connectors_service.connectors.sources.mysql.mysql import (
+from connectors.sources.mysql import (
     MySQLAdvancedRulesValidator,
     MySQLClient,
     MySqlDataSource,
